move functions logic to deprecated namespace
