+++ db/deprecated/functions/exceptions.py
