+++ db/deprecated/transforms/base.py
@@ -5,8 +5,8 @@ import itertools
 import sqlalchemy
 from sqlalchemy import select
 
-from db.functions.operations.apply import apply_db_function_by_id, apply_db_function_spec_as_filter
-from db.functions.packed import DistinctArrayAgg
+from db.deprecated.functions.operations.apply import apply_db_function_by_id, apply_db_function_spec_as_filter
+from db.deprecated.functions.packed import DistinctArrayAgg
 from db.records.operations import sort as rec_sort
