+++ db/tests/deprecated/functions/test_json.py
@@ -1,11 +1,11 @@
 import pytest
 from db.deprecated.utils import execute_pg_query
-from db.functions.base import ColumnName, Noop, JsonArrayContains
-from db.functions.packed import (
+from db.deprecated.functions.base import ColumnName, Noop, JsonArrayContains
+from db.deprecated.functions.packed import (
     JsonLengthEquals, JsonLengthGreaterThan, JsonLengthGreaterorEqual,
     JsonLengthLessThan, JsonLengthLessorEqual, JsonNotEmpty
 )
-from db.functions.operations.apply import apply_db_function_as_filter
+from db.deprecated.functions.operations.apply import apply_db_function_as_filter
 
 
 @pytest.mark.parametrize("main_db_function,literal_param,expected_count", [
