+++ mathesar/tests/rpc/exceptions/test_error_handler.py
@@ -4,7 +4,7 @@ from modernrpc.exceptions import RPCException
 from psycopg.errors import BadCopyFileFormat
 from django.core.exceptions import FieldDoesNotExist
 from mathesar.utils.permissions import BadInstallationTarget
-from db.functions.exceptions import UnknownDBFunctionID
+from db.deprecated.functions.exceptions import UnknownDBFunctionID
 from http.client import CannotSendRequest
 
 from mathesar.rpc.exceptions.handlers import handle_rpc_exceptions
